Paraguay geo 0.5 addition
There is a union command being applied and adding geo_0.5 to either dlt gives an error because there is a column number mismatch. If I add same lines of code to both the dlt (muncipal and cen), it generates the required geo_0.5 column, however there are additional lines of code then to make the column number equal for the union (pry_boost_silver) to take place. Would it make sense to create another dlt_gold_intermediate where post union of municipal and cen, we then create a geo_0.5 and move it to pry_boost_gold.